Epoch.
The basics work: NICK/USER, PINGs and JOIN/PART/PRIVMSG/NOTICE for channels. There is also a CAP stub. Rest? Noooope. Been working in this for a few days on and off. Works good enough for a basic testing server for my bots. Well, once I have modes and topics that is.